[REM-103] refactor: streamline process management and service startup
refactor:
- crawjud/__init__.py: removed Celery process management functions and updated service startup logic to use direct calls instead of threads.
- crawjud/core/config.py: improved process ID assignment logic for different process types.
- crawjud/manager/runner.py: restructured service startup to initialize all servers in a loop and added Celery worker and beat startup functions.
