+++ crawjud/__init__.py
@@ -1,28 +1,22 @@
 """Main module for the CrawJUD Package."""
 
 import asyncio
-from os import environ
 from pathlib import Path
-from platform import node
-from threading import Thread
 from time import sleep
 from typing import Callable, Dict
 
 import inquirer
 import rich
-from billiard.context import Process
-from celery.apps.beat import Beat
-from celery.apps.worker import Worker
 from clear import clear
 from socketio import AsyncServer
 from termcolor import colored
 from tqdm import tqdm
 
-from crawjud.core.config import StoreService, running_servers
 from crawjud.core import create_app
+from crawjud.core.config import StoreService as StoreService
+from crawjud.core.config import running_servers
 from crawjud.manager import HeadCrawjudManager
 from crawjud.types import app_name
-from crawjud.utils.gen_seed import worker_name_generator
 
 io = AsyncServer(
     async_mode="asgi",
@@ -31,64 +25,6 @@ io = AsyncServer(
     ping_timeout=10,
 )
 clear()
-
-
-def start_beat() -> None:
-    """Start the Celery beat scheduler."""
-    from crawjud.core import create_app
-
-    environ.update({"APPLICATION_APP": "beat"})
-
-    async def beat_start() -> None:
-        async with app.app_context():
-            beat = Beat(
-                app=celery,
-                scheduler="crawjud.utils.scheduler:DatabaseScheduler",
-                quiet=True,
-            )
-            beat.run()
-
-    app, _, celery = asyncio.run(create_app())
-    asyncio.run(beat_start())
-
-
-def start_worker() -> None:
-    """Start the Celery beat scheduler."""
-    from crawjud.core import create_app
-
-    app, _, celery = asyncio.run(create_app())
-    environ.update({"APPLICATION_APP": "worker"})
-
-    async def start_worker() -> None:
-        async with app.app_context():
-            worker_name = f"{worker_name_generator()}@{node()}"
-            worker = Worker(
-                app=celery,
-                hostname=worker_name,
-                task_events=True,
-                loglevel="INFO",
-                concurrency=50.0,
-                pool="threads",
-            )
-            worker = worker
-
-            try:
-                worker.start()
-
-            except Exception as e:
-                if isinstance(e, KeyboardInterrupt):
-                    worker.stop()
-
-                else:
-                    tqdm.write(
-                        colored(
-                            f"{colored('[ERROR]', 'red', attrs=['bold', 'blink'])} {e}",
-                            "red",
-                            attrs=["bold"],
-                        )
-                    )
-
-    asyncio.run(start_worker())
 
 
 class MasterApp(HeadCrawjudManager):
@@ -101,25 +37,10 @@ class MasterApp(HeadCrawjudManager):
         rich.print("[bold green]Starting application object...[/bold green]")
         self.app, self.asgi, self.celery = asyncio.run(create_app())
 
-        process_beat = Process(target=start_beat)
-        process_worker = Process(target=start_worker)
-
-        running_servers["Worker"] = StoreService(
-            process_name="Worker",
-            process_object=process_worker,
-            process_id=process_worker.pid,
-        )
-        running_servers["Beat"] = StoreService(
-            process_name="Beat",
-            process_object=process_beat,
-            process_id=process_beat.pid,
-        )
-        process_worker.start()
-        process_beat.start()
-
     def __init__(self) -> None:
         """Initialize the ASGI server."""
         self.boot_app()
+
         self.current_menu = self.main_menu
         self.current_menu_name = "Main Menu"
 
@@ -191,7 +112,7 @@ class MasterApp(HeadCrawjudManager):
             elif choice in options:
                 if choice == "Start Services":
                     call_obj = options.get(choice)
-                    Thread(target=call_obj, daemon=True).start()
+                    call_obj()
                 elif choice != "Start Services":
                     options.get(choice)()
